Butterfly/gestore_personale/requirements.txt
argparse
kafka-python
Flask
flask_restful
requests
bson
pymongo